move the files to V1 dir
